+++ cli/src/commands/config/set.rs
@@ -30,7 +30,7 @@ use crate::complete;
 use crate::config::parse_value_or_bare_string;
 use crate::ui::Ui;
 
-/// Update config file to set the given option to a given value.
+/// Update a config file to set the given option to a given value.
 #[derive(clap::Args, Clone, Debug)]
 pub struct ConfigSetArgs {
     #[arg(required = true, add = ArgValueCandidates::new(complete::leaf_config_keys))]